Adding loading the friends list on init
On init, we are not making a broadcast to say a friend joined the clique. The clique
then takes this message and sends back to everyone the new list of friends. Only issue
is the initial load is sending two broadcasts for the destination being set, but the geolocation
is empty.