Initial commit for the staleness
I'm able to set a flag when a person goes stale, but I'd ideally like
to notifiy all of the users this person stepped away and then
begin a countdown to mark them as being inactive. It'd be cool to change their marker
and have it countdown to show how much longer until the user times out.