Taking comments from PR and updating where the communication is sent
When initializing the visibility object, I pass in the communications object.

I removed the timeout for simplicity for now, once the user's session goes into
the background it will emit an event stating the user went state.

Removed the toast for when a user goes hidden.

Removed the updateMarker function and will defer the icon to the client.